mobile+backend: connect sign in to backend
Connect signing in to the backend. This feature involved several changes:

1. On the mobile side, after signing in via google we make a POST request
to the /sign_in endpoint on the backend with the user email in the JSON body

2. In the /sign_in endpoint we query the DB to see if the user exists and if they
do we return the user. Otherwise we add the user to the DB. This endpoint will error
out if no user is provided

3. The user table only consists of an id column and an email column. We should use
email to uniquely identify users. When writing any new backend endpoints use email
as the identifier unless you have access to the user id.