backend: add squad code
Added a squad code column on the backend as well as generating the
squad code - we do this by generating a short uuid & taking the
first 7 characters. This should largely be unique, but to
guarantee uniqueness we repeatedly query the db till we
generate a unique id. We probably won't have to query the db
more than once just to make sure that the uuid doesn't exist