backend: store user google tokens in db
Using the google authorization code, we fetch the user's access tokens
from google and store it in the db. We can then use these tokens to call
the google API for the user from the server. We store two tokens:
    * Access Token: This is the access token for the API. This has an
      expiry date
    * Refresh Token: When the access token expires, use this token to
      obtain a new token.

The current set up has some security issues in the event of data breaches:

    1) The tokens are stored in plaintext in our db. Now if anyone gets access
       to the tokens they can take actions for users if they have our project
       info. We should figure out either encrypting our db or at least encrypting
       these columns.

There's more info here on the best security practices for dealing with these tokens.
We don't have to complete these before sending our MVP to our close friends, but
they should be fixed before any sort of public beta:
https://developers.google.com/identity/protocols/oauth2/policies#secure-token-handling