mobile: fix datetimeinput tbd behavior
The component did not handle TBD behavior correctly. It would default
to displaying the current date in the text field if no date was
specified, but what we actually want is the text field to display
TBD and the datetimepicker to default to the current date on the
actual model itself not on the text field.

We fix this by decoupling the default date to be shown on the selector
and the default date displayed. This way, if a user doesn't specify
an initial date, the text will say TBD but the selector can still
default to the current time. Otherwise if they specify a date then
we will display that date.

If the user passes in a default date time, the text field will
display that, but if they don't then the text field will display
TBD.