mobile: force app to operate in light mode
Make the app operate in light mode, ignoring the user setting for
Dark mode for iOS. This fixes the datetimepicker issue.

The alternative would have been changing the appearance of the
datetimepicker based on dark mode. However, since the rest of our
app does not have a dark mode, I think it would have been weird
to only have the datetimepicker be dark mode and everything else
light, so instead I just decided to enforce light mode on our app.