mobile: automatically go to squads if signed in
On the login screen check if the user is signed in on the server.
If they are, automatically go to the squads screen. While we are
checking if the user is logged in, show a loading screen.

Changed the loading screen for the App screen as well to have a
consistent UI. Note we need to change the splash screen as well.