backend: setup FCM, FireBase SDK & notification for new event
Setup the FireBase SDK and FCM on the backend. The sdk requires
a credentials file which is not being checked into source code.
Instead I have it locally and have modified the deployment script
to fetch it from s3 to provision the server.

I also added a notification for when a new event is created. This
will send a notification to everyone in the squad except for the
creator.