mobile: fix bug with squad num on event page
When users join a squad which already has an event, the event details
page shows the wrong number for the total ppl in the squad. It calculates
it from the number of event responses. However, a new person joining the
squad does not have an event response so they are not counted in the
total amount.

This fixes the issue by fetching the total number of people in the squad
and using that for rsvp calculations. Also refactors a little bit to move
the get users call to  backend.tsx to make the code a little cleaner.