mobile: auto scroll view based on keyboard
Make a fix so that the keyboard doesn't cover up the focused text
box and instead the view automatically adjusts for the edit event page.

I tried to do the same with the multiline inputs here and the add event page
but it has issues. I just decided to at least make the multiline inputs
scrollable by adding a number of lines attribute.