mobile: replace useEffect with useFocusEffect
useEffect is only called the first time a page loads. However we
want components to call useEffect each time the page loads to get
updates. For example, we want the squad page to show a new event
after the event is added from the add event flow. To do this we need
to use useFocusEffect which is called every time a page loads.