backend: added scheduler & reminders to RSVP
Added APScheduler and set up a BackgroundScheduler that we can use
to schedule jobs at certain events. The jobs get stored in our db in
the apscheduler_jobs table. This table gets automatically created if
it doesn't exist when the app starts up.

I added a job to send a reminder to users to RSVP to an event if they
haven't accepted. The reminder is set 24 hours before the start of the
event and is updated if a user edits the event. This feature should
serve as an example of how to use the scheduler.