login: make setup flow sync
Force the setup flow to execute synchronously using async/await. This might fix a potential bug with push notifications. Since we check if the user is signed in while we also try to set the device token, we might not end up setting the device token properly. We force the flow to be synchronous to set the device token. We also add in a call to the backend to set the device token for the user.