backend: fix event image url column to be text
Change the image url column to be a regular text column without a specific size. This is to stop Flask from trying to migrate the image url column to Text(65536) over and over again.

Tested by rerunning flask db migrate after the upgrade and making sure that it didn't try to migrate the column again.

Also made sure that adding an image to an event still worked.