mobile+backend: fix push notifications issue
There was a really subtle issue here in fetching the device token from firebase and assigning it to the deviceToken state. This deviceToken state variable gets sent to the backend and updates the users deviceToken.

However in React, the useState update hook doesn't actually update the variable immediately. The variable is updated on the next rerender.So even though we've called setDeviceToken, the variable doesn't actually get updated till the next rerender. This async behavior is unavoidable; it can't be dealt with using async/await as previously thought.

The fix is pretty simple - we just set the deviceToken variable to be a simple class variable instead of being part of the component state. The broader lesson here is unless a variable is being used for the render, it shouldn't be part of the state - we should just make it a simple constant.

I also added a fix to the backend to make sure we only update the device token if the one that's passed in is not null to avoid overwriting valid device tokens.