mobile: redesign edit event screen
Redesign the edit event screen. The edit event screen now is just the same as the Add Event Screen but with the details prefilled in.

We leverage the same component for both Adding and Editing events, & pass in a isEditView variable to deal with the minor differences between the two components.