backend: fix migrations for event image url column
We cannot apply these migrations for the image url column in the event table. Our data is too big and so the intermediate migrations to convert these to TEXT columns fails.
I've just made them pass for now since it's difficult to actually delete a migration. I also have a migration to upgrade it to a LONGTEXT column bc I
misunderstood the error at first and now this migration has been applied to the database.